+++ dpnp/backend/extensions/window/window_py.cpp
@@ -1,5 +1,5 @@
 //*****************************************************************************
-// Copyright (c) 2024-2025, Intel Corporation
+// Copyright (c) 2025, Intel Corporation
 // All rights reserved.
 //
 // Redistribution and use in source and binary forms, with or without
@@ -23,21 +23,15 @@
 // THE POSSIBILITY OF SUCH DAMAGE.
 //*****************************************************************************
 //
-// This file defines functions of dpnp.backend._statistics_impl extensions
+// This file defines functions of dpnp.backend._window_impl extensions
 //
 //*****************************************************************************
 
 #include <pybind11/pybind11.h>
 
-#include "bincount.hpp"
-#include "histogram.hpp"
-#include "histogramdd.hpp"
-#include "sliding_dot_product1d.hpp"
+#include "hamming.hpp"
 
-PYBIND11_MODULE(_statistics_impl, m)
+PYBIND11_MODULE(_window_impl, m)
 {
-    statistics::histogram::populate_bincount(m);
-    statistics::histogram::populate_histogram(m);
-    statistics::sliding_window1d::populate_sliding_dot_product1d(m);
-    statistics::histogram::populate_histogramdd(m);
+    dpnp::extensions::window::init_hamming(m);
 }