+++ dpnp/dpnp_iface.py
@@ -106,6 +106,8 @@ from dpnp.dpnp_iface_statistics import *
 from dpnp.dpnp_iface_statistics import __all__ as __all__statistics
 from dpnp.dpnp_iface_trigonometric import *
 from dpnp.dpnp_iface_trigonometric import __all__ as __all__trigonometric
+from dpnp.dpnp_iface_window import *
+from dpnp.dpnp_iface_window import __all__ as __all__window
 
 # pylint: disable=no-name-in-module
 from .dpnp_utils import (
@@ -130,6 +132,7 @@ __all__ += __all__searching
 __all__ += __all__sorting
 __all__ += __all__statistics
 __all__ += __all__trigonometric
+__all__ += __all__window
 
 
 def are_same_logical_tensors(ar1, ar2):